# Slot We want to display a slot. A slot is a empty circle with surrounded with a blue square.
## TODOs
 - [ ] Create a blue squares
 - [ ] Inside the blue square leave a white circle

## Details
We use css to add styling to our html elements. We want to create two elements. The first element we create a `div` where we create a `background-color` of `blue`. Inside that element, we want to create another html element with a `transparent` `background-color`. It has equal `height` and `width` with a `border-radius` of `50%` to make it round.

## Further Reading
 - [https://www.google.com/search?q=mozilla+css+background+color+transparent](mozilla css background color transparent) for mozilla documentation on background color
 - [https://www.google.com/search?q=mozilla+css+round+element](mozilla css round element) for mozilla documentation on border-radius
 - [https://www.google.com/search?q=mozilla+margin](mozilla margin) for mozilla documentation on css margins